requirements.txt
openai
PyPDF2
spire.PDF
pytesseract
img2table